+++ .gitignore
@@ -1,4 +1,4 @@
 .idea
-.out
+out
 
 DSeccion10.iml
