--- src/MainForm.java
@@ -1,8 +1,0 @@
-import javax.swing.*;
-
-public class MainForm {
-    private JTextPane pricePane;
-    private JTextPane taxPane;
-    private JTextPane totalPane;
-    private JButton buttonCalculate;
-}
